+++ src/decorators/componentDisplayFlex.decorator.ts
@@ -1,0 +1,11 @@
+import {ComponentDisplay} from './componentDisplay.decorator';
+
+/**
+ * Applies css display flex style to host of component, requires at least empty style on Component
+ * 
+ * @experimental
+ */
+export function ComponentDisplayFlex(): ClassDecorator
+{
+    return ComponentDisplay('flex');
+}