+++ src/decorators/scrollableContent.decorator.ts
@@ -1,0 +1,11 @@
+import {ComponentHostStyle} from './componentHostStyle.decorator';
+
+/**
+ * Applies css overflow-y scroll style to host of component, requires at least empty style on Component
+ * 
+ * @experimental
+ */
+export function ScrollableContent(): ClassDecorator
+{
+    return ComponentHostStyle('overflow-y: scroll;');
+}