+++ src/modules/progressIndicator/modules/progressIndicator.module.ts
@@ -1,7 +1,7 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
 
-import {CommonLocalizeModule} from '../../commonLocalize';
+import {CommonLocalizeModule} from '../../commonLocalize.module';
 import {ProgressIndicatorComponent} from '../components/progressIndicator/progressIndicator.component';
 import {ProgressOverlayDirective} from '../directives/progressOverlay/progressOverlay.directive';